+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- No changes
+### Changed
+- Modal close button is now the default iOS 'stop' icon instead of a custom one
+
 
 ## [0.3.3] - 2016-03-07
 ### Fixed